react-app/backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
Pillow
python-multipart
